refactor!: move enums from library/CELib to separate classes in an enums folder
