+++ src/main/java/com/cewar/enums/Archetype.java
@@ -1,0 +1,16 @@
+package com.cewar.enums;
+
+/**
+ * Archetype(s) of a card
+ * 
+ * <p> Consists of both series archetypes and functional archetypes
+ */
+public enum Archetype {
+    // Traditional Archetypes
+    STAPLE, FEATHERY, SKELL, BUGGLE, AQUATIC, MATERIAL, TAROT, CODED, SLIME, ARACHNIC, LIZAR,
+    HUMAN, FORTUNE, DRAGON,
+                        
+    // Functional Archetypes
+    INFERIOR, SUPERIOR, ANIMATED, ANCHORED, CONJURED, DEPENDENT, IMPASSABLE, ACCESSIBLE, 
+    DIRECTIONAL, AFLOAT, EXTRA
+}
