+++ src/main/java/com/cewar/enums/Attribute.java
@@ -1,0 +1,16 @@
+package com.cewar.enums;
+
+/**
+ * Attribute(s) of a card
+ * 
+ * <p> These are displayed before the name, except for GOD, which is displayed in place of the rarity symbol.
+ */
+public enum Attribute {
+    C1,
+    C2,
+    C3,
+    C4,
+    STAGE,
+    EQUIP,
+    GOD 
+}
