+++ src/main/java/com/cewar/enums/Location.java
@@ -1,0 +1,13 @@
+package com.cewar.enums;
+
+/**
+ * Used by {@link com.cewar.game.GameCard GameCards} to refer to their location in a game.
+ */
+public enum Location {
+    HAND,
+    DECK,
+    FIELD,
+    GRAVE,
+    DUMP,
+    EXTRA
+}
