+++ src/main/java/com/cewar/enums/Rarity.java
@@ -1,0 +1,12 @@
+package com.cewar.enums;
+
+/**
+ * Rarity of a card
+ */
+public enum Rarity {
+    COMMON,
+    UNCOMMON,
+    RARE,
+    ULTRA_RARE,
+    NONE
+}
