+++ src/main/java/com/cewar/enums/Type.java
@@ -1,0 +1,14 @@
+package com.cewar.enums;
+
+/**
+ * Type (color) of a card.
+ */
+public enum Type {
+    CREATURE,
+    ACTION,
+    BUILDING,
+    MATERIAL,
+    GOD,
+    REPLICA,
+    OTHER
+}
